Enclave ID must be 8 bytes
eid had been defined as 4 bytes where SM writes 8 bytes to the pointer.
This overwrites the close_on_pexit variable

+++ keystone.h
@@ -80,7 +80,7 @@ struct utm {
 
 struct enclave
 {
-  unsigned int eid;
+  unsigned long eid;
   int close_on_pexit;
   struct utm* utm;
   struct epm* epm;
